[hotfix] cache related fixed.
